Cleans up RBM constructor
I moved the code for calculating rating pmf's into the RBM constructor,
for ease of use, and modified the constructor to take the number of
users, movies, and hidden features as constructor parameters.  Also
tracked my super simplistic (for now) test binary.  And for the sake of
readability & consistency, I renamed the SVD++ test binary (as well as all
references to it).