Shuffling in svdpp_test; larger training set.
Note that the SVDPP class assumes that the "data" (passed into the
"train" method) is shuffled beforehand (to ensure stochasticity). I
added a function in svdpp_test that can quickly shuffle the "rows" in a
vector, so it might be useful to use that before converting vectors into
Armadillo matrices. Lastly, svdpp_test now uses all of the training set
(but not "probe").