Added RBM helpers, edge cases
I encapsulated the code for computing & caching an indicator matrix from
a data matrix.  Also wrote a simple binary search template for use in
the edge case where only some indicator matrices are missing, which is
now fully supported.  Finally, I wrote a helper for converting a
probability matrix into a Bernoulli matrix; elements in the matrix are
set to 1 if they are greater than a uniform random number in the range
[0, 1], and 0 otherwise