Adds instrumentation to RBM::train, bugfix
I added instrumentation the the RBM::train method to identify the mostly
costly operations in the algorithm; it turns out that zeroing the loop
arrays after each user takes ~.0006 seconds, and the time for all other
operations is cumulatively about 3+ orders of magnitude lower.  Also
fixed a segfault that was occurring because I accidently incremented a
double pointer (pointer array) instead of one its elements;  all
training arrays are now const pointers (as opposed to pointers to
const's).