Undid residual-caching changes to global effects.
These were really slow, and Armadillo's sparse matrix batch insertion
constructors didn't seem to work for some reason (in particular, they
only worked when a subset of the 100 million+ ratings were used for
insertion). I've manually undone the earlier changes, and the
performance of global effects seems to be the same as it was before (a
probe RMSE of 0.966201 with 10 global effects).