kNN + 10 GE is working; 1.07% above water.
In particular, the quiz RMSE for this combination is 0.94117. Note that
both of these algorithms were trained on the entire dataset. A
corresponding entry has been added to the "good_predictions" folder too.

I had to make a couple of changes to how ComboAlgorithms work, but you
can refer to knn_on_globals.cc to see the current procedure that's used
to train combo algorithms. We might want to encapsulate this procedure
even more, since it's fairly general.