Added 60-factor Time-SVD++ prediction.
This run was carried out on the entire dataset, with 60 factors, 30 time
bins, 40 iterations, and the same LRs/RFs as in TIMESVDPP_QUAL_7.749.
The resulting qual predictions had a QRMSE of 0.87911 (7.598% above
water). Note that this run *did* use userFactorTime.

I also added some scoping to the sparse matrix initialization in
Time-SVD++, so that it's a little bit more memory efficient.