Removed 3 highly correlated predictors.
If there were pairs of predictors with very high correlations (> around
0.96 or 0.97) and very similar QRMSEs, then I removed the predictor with
the worse QRMSE in that pair. So far, I've removed three predictors from
the quiz blend, and moved them to "old_predictors". This made the quiz
blend QRMSE become 0.86963 instead of 0.86962, but should improve
performance on the test set.