Removed 400-factor RBM predictor from blend.
This was really highly correlated with a kNN-on-RBM prediction that was
carried out on the results of this RBM prediction, so I've moved the RBM
prediction to old_predictors. The QRMSE of the blend is unchanged,
according to the scoreboard.

I also modified quiz_blend.py to bring out the correlations of > 0.85
more clearly (black => 0.85 correlation now).