Removed 1 SVD predictor.
The 1000-factor 80-epoch SVD predictor was too highly correlated with
the 2000-factor one, so I removed the 1000-factor one and started
generating a new 1000-factor version (which will be run for fewer epochs
and won't be trained on probe). The QRMSE of the blend did not change
when I did this.

I also made src/quiz_blend.py parse predictors in alphabetical order
(from the data/quiz_blend folder), instead of doing that in "last
modified" order. This makes the heatmap more visually useful.