Added two predictors; now 8.683% above water.
These predictors came from a 1000-factor SVD run and a 750-factor SVD++
run, both of which were **not** trained on probe. The SVD really
improved the QRMSE, and the SVD++ only helped a little (since it was
fairly correlated with a previous SVD++ non-probe run, whereas the SVD
predictions were not so correlated).