Added quiz bin blending.
This is currently set up to automatically submit to the scoreboard with
our team ID, and managed to rack up several 100 (in fact > 1000)
submissions today. There's also caching to prevent unnecessary
recomputing, but it's important to remove the caches (in
data/quiz_bin_blend)

Note that the quiz_bin_blend script must be run with Python 2, and has
the same usage as the regular quiz_blend script. Also, note that the
generate_bins helper script (which also uses Python 2) should give all
of the prerequisites that the script needs (which is just the bin
demarcation file "bins.dta").