Added final blend predictions: 8.756% AW.
This came from a 15-bin quiz blend with evenly split bins. I also set
lambda = 0, which is a bit justifiable since the datasets are so small
that large lambdas are problematic. However, this might have also led to
overfitting on the quiz set. We didn't have time to test different
parameters, bin boundaries, lambda values, etc, so eh.